+++ libs/zard/src/lib/shared/components/toggle/demo/outline.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 
 import { NgIcon, provideIcons } from '@ng-icons/core';
-import { lucideBold } from '@ng-icons/lucide';
+import { lucideBold, lucideItalic } from '@ng-icons/lucide';
 
 import { ZardToggleComponent } from '../toggle.component';
 
@@ -9,10 +9,17 @@ import { ZardToggleComponent } from '../toggle.component';
   selector: 'z-demo-toggle-outline',
   imports: [ZardToggleComponent, NgIcon],
   template: `
-    <z-toggle aria-label="Toggle outline" zType="outline">
-      <ng-icon name="lucideBold" />
-    </z-toggle>
+    <div class="flex items-center gap-2">
+      <z-toggle zAriaLabel="Toggle italic" zType="outline">
+        <ng-icon name="lucideItalic" />
+        Italic
+      </z-toggle>
+      <z-toggle zAriaLabel="Toggle bold" zType="outline">
+        <ng-icon name="lucideBold" />
+        Bold
+      </z-toggle>
+    </div>
   `,
-  viewProviders: [provideIcons({ lucideBold })],
+  viewProviders: [provideIcons({ lucideBold, lucideItalic })],
 })
 export class ZardDemoToggleOutlineComponent {}